Outputs a pirate plot (RDI)

Usage

pirateye(
  data,
  colour_condition = NULL,
  x_condition = "variable",
  facet_condition = NULL,
  facet_scales = "fixed",
  pid_average = F,
  dv,
  reorder = F,
  pid = "pid",
  dodgewidth = 0.8,
  plot_condition = NULL,
  cond = NULL,
  cond2 = NULL,
  facetby = NULL,
  ylim = NULL,
  xlim = NULL,
  w = NULL,
  h = 6,
  title = NULL,
  outp = "analysis",
  cols = NULL,
  pred_line = F,
  error_bar_data = NULL,
  ypercent = F,
  pred = NULL,
  pred_means = NULL,
  pred_bar = T,
  xlabs = NULL,
  xlabpos = 0.7,
  error_data = NULL,
  cflip = F,
  norm = F,
  bars = F,
  violin = T,
  dots = T,
  splitV = F,
  svw = 1,
  dot_h_jitter = 0,
  line = F,
  error_bars = T,
  useall = F,
  legend = T,
  title_overide = F,
  combine_plots = list(),
  combine_position = "right",
  elementinc = NULL,
  type = NULL,
  ...
)

Arguments

data

data with one person per line, excluding rows with use=0

colour_condition

colour split

x_condition

x axis split (if not, specified colour condition used for x axis too)

facet_condition

for faceting

pid_average

plot an average of each participant's dv over the named conditions

dv

name of single dv column, or multiple columns, in which case they will be split by x_condition unless colour or facet condition set to 'variable'

pid

whats the name of col that identifies individuals

plot_condition

instead of passing individually, you can give a vector of up to 3

cols

specify the colours to use, can be a set of colours, or of condition_level to colour

ypercent

Convert y scale to

xlabposHow high vertically should they be, as proportion of plot height

error_datadistribution for mean, eg from Bayes analysis, to replace SE

cflipflip to horizontal plot

normnormalise / z-score values for comparison across scales

useallignore the use column and plot all rows

typeshortcuts: m=just error bars, b=just bars

xlabDo we have labels to go across x axis, such as post hoc pvalues or MPEs

Defaults to violin plot with error bars and dots, but elements such as dots, bars, violin, error_bars can be turned on or off